experiments/elevenlabs_verification/pyproject.toml
[project]
name = "elevenlabs-verification"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "gradio>=6.8.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "playwright>=1.58.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "scipy>=1.17.1",
]
